@@ -4,7 +4,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-    sdk: '>=2.6.0 <3.0.0'
+    sdk: '>=3.4.3 <4.0.0'
 
 dependencies:
     flutter:
@@ -12,23 +12,26 @@ dependencies:
     flutter_localizations:
         sdk: flutter
     cupertino_icons: ^1.0.0
-    sqflite: ^1.2.2
+    sqflite: ^2.3.3+1
     path: ^1.6.4
-    provider: ^4.1.2
-    flutter_svg: ^0.18.1
-    package_info: ^0.4.0+14
-    html: ^0.14.0+3
-    shared_preferences: ^0.5.7
-    flutter_colorpicker: ^0.3.4
-    intl: ^0.16.1
-    get_it: ^4.0.2
+    provider: ^6.1.2
+    flutter_svg: ^2.0.10+1
+    package_info: ^2.0.2
+    html: ^0.15.4
+    shared_preferences: ^2.3.1
+    flutter_colorpicker: ^1.1.0
+    intl: ^0.19.0
+    get_it: ^7.7.0
+    flutter_widget_from_html: ^0.15.1
+    html_unescape: ^2.0.0
+    catholic_liturgy: ^0.0.17
 
 dev_dependencies:
     flutter_test:
         sdk: flutter
-    enum_to_string: ^1.0.9
-    mockito: ^4.1.1
-    csv: ^4.0.3
+    enum_to_string: ^2.0.1
+    mockito: ^5.4.4
+    csv: ^6.0.0
 
 flutter:
     uses-material-design: true
