@@ -11,7 +11,7 @@ import 'components/bible_books_page.dart';
 class BibleBooks extends StatelessWidget {
   final TestamentModel testament;
 
-  const BibleBooks({this.testament, Key key}) : super(key: key);
+  const BibleBooks({required this.testament, Key? key}) : super(key: key);
 
   static oldTestament(BuildContext context) {
     return TestamentModel(
