@@ -5,9 +5,9 @@ import 'package:get_it/get_it.dart';
 
 class BibleBooksState with ChangeNotifier {
   final service = GetIt.instance;
-  List<CategoryModel> _model;
+  List<CategoryModel>? _model;
 
-  List<CategoryModel> get model {
+  List<CategoryModel>? get model {
     return _model;
   }
 
