@@ -11,7 +11,7 @@ import 'bible_html_text.dart';
 class BibleTextPage extends StatelessWidget {
   final ChapterModel chapter;
 
-  BibleTextPage({@required this.chapter});
+  BibleTextPage({required this.chapter});
 
   @override
   Widget build(BuildContext context) {
