@@ -4,7 +4,7 @@ import 'package:bibbia_cattolica/common/theme.dart';
 class PreferenceTitle extends StatelessWidget {
   final String title;
 
-  const PreferenceTitle({@required this.title, Key key}) : super(key: key);
+  const PreferenceTitle({required this.title, Key key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
