@@ -11,25 +11,21 @@ class AppBarTitle extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    if (model.subtitle == null) {
-      return Text(model.title);
-    } else {
-      return Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              model.title,
-              style: TextStyle(fontSize: 18.0),
+    return Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            model.title,
+            style: TextStyle(fontSize: 18.0),
+          ),
+          Padding(
+            padding: EdgeInsets.only(top: 4),
+            child: Text(
+              model.subtitle,
+              style: TextStyle(fontSize: 16.0),
             ),
-            Padding(
-              padding: EdgeInsets.only(top: 4),
-              child: Text(
-                model.subtitle,
-                style: TextStyle(fontSize: 16.0),
-              ),
-            ),
-          ]);
+          ),
+        ]);
     }
-  }
 }
