@@ -1,5 +1,5 @@
 class MenuChoice {
-  const MenuChoice({this.title, this.onTap});
+  const MenuChoice({required this.title, required this.onTap});
 
   final String title;
   final Function onTap;
